requirements.txt
fastapi
uvicorn
openai-agents
requests
python-dotenv
wikipedia